% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_get_archive_rule}
\alias{accessanalyzer_get_archive_rule}
\title{Retrieves information about an archive rule}
\usage{
accessanalyzer_get_archive_rule(analyzerName, ruleName)
}
\arguments{
\item{analyzerName}{[required] The name of the analyzer to retrieve rules from.}

\item{ruleName}{[required] The name of the rule to retrieve.}
}
\description{
Retrieves information about an archive rule.
}
\details{
To learn about filter keys that you can use to create an archive rule,
see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/access-analyzer-reference-filter-keys.html}{Access Analyzer filter keys}
in the \strong{IAM User Guide}.
}
\section{Request syntax}{
\preformatted{svc$get_archive_rule(
  analyzerName = "string",
  ruleName = "string"
)
}
}

\keyword{internal}
