% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_get_finding}
\alias{accessanalyzer_get_finding}
\title{Retrieves information about the specified finding}
\usage{
accessanalyzer_get_finding(analyzerArn, id)
}
\arguments{
\item{analyzerArn}{[required] The ARN of the analyzer that generated the finding.}

\item{id}{[required] The ID of the finding to retrieve.}
}
\description{
Retrieves information about the specified finding.
}
\section{Request syntax}{
\preformatted{svc$get_finding(
  analyzerArn = "string",
  id = "string"
)
}
}

\keyword{internal}
