% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_list_analyzed_resources}
\alias{accessanalyzer_list_analyzed_resources}
\title{Retrieves a list of resources of the specified type that have been
analyzed by the specified analyzer}
\usage{
accessanalyzer_list_analyzed_resources(analyzerArn, maxResults,
  nextToken, resourceType)
}
\arguments{
\item{analyzerArn}{[required] The ARN of the analyzer to retrieve a list of analyzed resources from.}

\item{maxResults}{The maximum number of results to return in the response.}

\item{nextToken}{A token used for pagination of results returned.}

\item{resourceType}{The type of resource.}
}
\description{
Retrieves a list of resources of the specified type that have been
analyzed by the specified analyzer..
}
\section{Request syntax}{
\preformatted{svc$list_analyzed_resources(
  analyzerArn = "string",
  maxResults = 123,
  nextToken = "string",
  resourceType = "AWS::S3::Bucket"|"AWS::IAM::Role"|"AWS::SQS::Queue"|"AWS::Lambda::Function"|"AWS::Lambda::LayerVersion"|"AWS::KMS::Key"
)
}
}

\keyword{internal}
