% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_list_archive_rules}
\alias{accessanalyzer_list_archive_rules}
\title{Retrieves a list of archive rules created for the specified analyzer}
\usage{
accessanalyzer_list_archive_rules(analyzerName, maxResults, nextToken)
}
\arguments{
\item{analyzerName}{[required] The name of the analyzer to retrieve rules from.}

\item{maxResults}{The maximum number of results to return in the request.}

\item{nextToken}{A token used for pagination of results returned.}
}
\description{
Retrieves a list of archive rules created for the specified analyzer.
}
\section{Request syntax}{
\preformatted{svc$list_archive_rules(
  analyzerName = "string",
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
