% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_create_control}
\alias{auditmanager_create_control}
\title{Creates a new custom control in AWS Audit Manager}
\usage{
auditmanager_create_control(name, description, testingInformation,
  actionPlanTitle, actionPlanInstructions, controlMappingSources, tags)
}
\arguments{
\item{name}{[required] The name of the control.}

\item{description}{The description of the control.}

\item{testingInformation}{The steps to follow to determine if the control has been satisfied.}

\item{actionPlanTitle}{The title of the action plan for remediating the control.}

\item{actionPlanInstructions}{The recommended actions to carry out if the control is not fulfilled.}

\item{controlMappingSources}{[required] The data source that determines from where AWS Audit Manager collects
evidence for the control.}

\item{tags}{The tags associated with the control.}
}
\description{
Creates a new custom control in AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$create_control(
  name = "string",
  description = "string",
  testingInformation = "string",
  actionPlanTitle = "string",
  actionPlanInstructions = "string",
  controlMappingSources = list(
    list(
      sourceName = "string",
      sourceDescription = "string",
      sourceSetUpOption = "System_Controls_Mapping"|"Procedural_Controls_Mapping",
      sourceType = "AWS_Cloudtrail"|"AWS_Config"|"AWS_Security_Hub"|"AWS_API_Call"|"MANUAL",
      sourceKeyword = list(
        keywordInputType = "SELECT_FROM_LIST",
        keywordValue = "string"
      ),
      sourceFrequency = "DAILY"|"WEEKLY"|"MONTHLY",
      troubleshootingText = "string"
    )
  ),
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
