% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_disassociate_assessment_report_evidence_folder}
\alias{auditmanager_disassociate_assessment_report_evidence_folder}
\title{Disassociates an evidence folder from the specified assessment report in
AWS Audit Manager}
\usage{
auditmanager_disassociate_assessment_report_evidence_folder(
  assessmentId, evidenceFolderId)
}
\arguments{
\item{assessmentId}{[required] The identifier for the specified assessment.}

\item{evidenceFolderId}{[required] The identifier for the folder in which evidence is stored.}
}
\description{
Disassociates an evidence folder from the specified assessment report in
AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$disassociate_assessment_report_evidence_folder(
  assessmentId = "string",
  evidenceFolderId = "string"
)
}
}

\keyword{internal}
