% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_assessment_report_url}
\alias{auditmanager_get_assessment_report_url}
\title{Returns the URL of a specified assessment report in AWS Audit Manager}
\usage{
auditmanager_get_assessment_report_url(assessmentReportId, assessmentId)
}
\arguments{
\item{assessmentReportId}{[required] The identifier for the assessment report.}

\item{assessmentId}{[required] The identifier for the specified assessment.}
}
\description{
Returns the URL of a specified assessment report in AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$get_assessment_report_url(
  assessmentReportId = "string",
  assessmentId = "string"
)
}
}

\keyword{internal}
