% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_change_logs}
\alias{auditmanager_get_change_logs}
\title{Returns a list of changelogs from AWS Audit Manager}
\usage{
auditmanager_get_change_logs(assessmentId, controlSetId, controlId,
  nextToken, maxResults)
}
\arguments{
\item{assessmentId}{[required] The identifier for the specified assessment.}

\item{controlSetId}{The identifier for the specified control set.}

\item{controlId}{The identifier for the specified control.}

\item{nextToken}{The pagination token used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results per page, or per API request
call.}
}
\description{
Returns a list of changelogs from AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$get_change_logs(
  assessmentId = "string",
  controlSetId = "string",
  controlId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
