% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_evidence_by_evidence_folder}
\alias{auditmanager_get_evidence_by_evidence_folder}
\title{Returns all evidence from a specified evidence folder in AWS Audit
Manager}
\usage{
auditmanager_get_evidence_by_evidence_folder(assessmentId, controlSetId,
  evidenceFolderId, nextToken, maxResults)
}
\arguments{
\item{assessmentId}{[required] The identifier for the specified assessment.}

\item{controlSetId}{[required] The identifier for the control set.}

\item{evidenceFolderId}{[required] The unique identifier for the folder in which the evidence is stored.}

\item{nextToken}{The pagination token used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results per page, or per API request
call.}
}
\description{
Returns all evidence from a specified evidence folder in AWS Audit
Manager.
}
\section{Request syntax}{
\preformatted{svc$get_evidence_by_evidence_folder(
  assessmentId = "string",
  controlSetId = "string",
  evidenceFolderId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
