% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_evidence_folders_by_assessment_control}
\alias{auditmanager_get_evidence_folders_by_assessment_control}
\title{Returns a list of evidence folders associated with a specified control
of an assessment in AWS Audit Manager}
\usage{
auditmanager_get_evidence_folders_by_assessment_control(assessmentId,
  controlSetId, controlId, nextToken, maxResults)
}
\arguments{
\item{assessmentId}{[required] The identifier for the specified assessment.}

\item{controlSetId}{[required] The identifier for the specified control set.}

\item{controlId}{[required] The identifier for the specified control.}

\item{nextToken}{The pagination token used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results per page, or per API request
call.}
}
\description{
Returns a list of evidence folders associated with a specified control
of an assessment in AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$get_evidence_folders_by_assessment_control(
  assessmentId = "string",
  controlSetId = "string",
  controlId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
