% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_organization_admin_account}
\alias{auditmanager_get_organization_admin_account}
\title{Returns the name of the delegated AWS administrator account for the AWS
organization}
\usage{
auditmanager_get_organization_admin_account()
}
\description{
Returns the name of the delegated AWS administrator account for the AWS
organization.
}
\section{Request syntax}{
\preformatted{svc$get_organization_admin_account()
}
}

\keyword{internal}
