% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_list_assessment_frameworks}
\alias{auditmanager_list_assessment_frameworks}
\title{Returns a list of the frameworks available in the AWS Audit Manager
framework library}
\usage{
auditmanager_list_assessment_frameworks(frameworkType, nextToken,
  maxResults)
}
\arguments{
\item{frameworkType}{[required] The type of framework, such as standard or custom.}

\item{nextToken}{The pagination token used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results per page, or per API request
call.}
}
\description{
Returns a list of the frameworks available in the AWS Audit Manager
framework library.
}
\section{Request syntax}{
\preformatted{svc$list_assessment_frameworks(
  frameworkType = "Standard"|"Custom",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
