% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_list_tags_for_resource}
\alias{auditmanager_list_tags_for_resource}
\title{Returns a list of tags for the specified resource in AWS Audit Manager}
\usage{
auditmanager_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the specified resource.}
}
\description{
Returns a list of tags for the specified resource in AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  resourceArn = "string"
)
}
}

\keyword{internal}
