% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_update_assessment}
\alias{auditmanager_update_assessment}
\title{Edits an AWS Audit Manager assessment}
\usage{
auditmanager_update_assessment(assessmentId, assessmentName,
  assessmentDescription, scope, assessmentReportsDestination, roles)
}
\arguments{
\item{assessmentId}{[required] The identifier for the specified assessment.}

\item{assessmentName}{The name of the specified assessment to be updated.}

\item{assessmentDescription}{The description of the specified assessment.}

\item{scope}{[required] The scope of the specified assessment.}

\item{assessmentReportsDestination}{The assessment report storage destination for the specified assessment
that is being updated.}

\item{roles}{The list of roles for the specified assessment.}
}
\description{
Edits an AWS Audit Manager assessment.
}
\section{Request syntax}{
\preformatted{svc$update_assessment(
  assessmentId = "string",
  assessmentName = "string",
  assessmentDescription = "string",
  scope = list(
    awsAccounts = list(
      list(
        id = "string",
        emailAddress = "string",
        name = "string"
      )
    ),
    awsServices = list(
      list(
        serviceName = "string"
      )
    )
  ),
  assessmentReportsDestination = list(
    destinationType = "S3",
    destination = "string"
  ),
  roles = list(
    list(
      roleType = "PROCESS_OWNER"|"RESOURCE_OWNER",
      roleArn = "string"
    )
  )
)
}
}

\keyword{internal}
