% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_update_assessment_control}
\alias{auditmanager_update_assessment_control}
\title{Updates a control within an assessment in AWS Audit Manager}
\usage{
auditmanager_update_assessment_control(assessmentId, controlSetId,
  controlId, controlStatus, commentBody)
}
\arguments{
\item{assessmentId}{[required] The identifier for the specified assessment.}

\item{controlSetId}{[required] The identifier for the specified control set.}

\item{controlId}{[required] The identifier for the specified control.}

\item{controlStatus}{The status of the specified control.}

\item{commentBody}{The comment body text for the specified control.}
}
\description{
Updates a control within an assessment in AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$update_assessment_control(
  assessmentId = "string",
  controlSetId = "string",
  controlId = "string",
  controlStatus = "UNDER_REVIEW"|"REVIEWED"|"INACTIVE",
  commentBody = "string"
)
}
}

\keyword{internal}
