% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_update_assessment_status}
\alias{auditmanager_update_assessment_status}
\title{Updates the status of an assessment in AWS Audit Manager}
\usage{
auditmanager_update_assessment_status(assessmentId, status)
}
\arguments{
\item{assessmentId}{[required] The identifier for the specified assessment.}

\item{status}{[required] The current status of the specified assessment.}
}
\description{
Updates the status of an assessment in AWS Audit Manager.
}
\section{Request syntax}{
\preformatted{svc$update_assessment_status(
  assessmentId = "string",
  status = "ACTIVE"|"INACTIVE"
)
}
}

\keyword{internal}
