% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsm_operations.R
\name{cloudhsm_delete_hapg}
\alias{cloudhsm_delete_hapg}
\title{This is documentation for AWS CloudHSM Classic}
\usage{
cloudhsm_delete_hapg(HapgArn)
}
\arguments{
\item{HapgArn}{[required] The ARN of the high-availability partition group to delete.}
}
\description{
This is documentation for \strong{AWS CloudHSM Classic}. For more
information, see \href{https://aws.amazon.com/cloudhsm/faqs/}{AWS CloudHSM Classic FAQs}, the \href{https://docs.aws.amazon.com/cloudhsm/classic/userguide/}{AWS CloudHSM Classic User Guide},
and the \href{https://docs.aws.amazon.com/cloudhsm/classic/APIReference/}{AWS CloudHSM Classic API Reference}.
}
\details{
\strong{For information about the current version of AWS CloudHSM}, see \href{https://aws.amazon.com/cloudhsm/}{AWS CloudHSM}, the \href{https://docs.aws.amazon.com/cloudhsm/latest/userguide/}{AWS CloudHSM User Guide}, and the
\href{https://docs.aws.amazon.com/cloudhsm/latest/APIReference/}{AWS CloudHSM API Reference}.

Deletes a high-availability partition group.
}
\section{Request syntax}{
\preformatted{svc$delete_hapg(
  HapgArn = "string"
)
}
}

\keyword{internal}
