% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsmv2_operations.R
\name{cloudhsmv2_modify_backup_attributes}
\alias{cloudhsmv2_modify_backup_attributes}
\title{Modifies attributes for AWS CloudHSM backup}
\usage{
cloudhsmv2_modify_backup_attributes(BackupId, NeverExpires)
}
\arguments{
\item{BackupId}{[required] The identifier (ID) of the backup to modify. To find the ID of a backup,
use the DescribeBackups operation.}

\item{NeverExpires}{[required] Specifies whether the service should exempt a backup from the retention
policy for the cluster. \code{True} exempts a backup from the retention
policy. \code{False} means the service applies the backup retention policy
defined at the cluster.}
}
\description{
Modifies attributes for AWS CloudHSM backup.
}
\section{Request syntax}{
\preformatted{svc$modify_backup_attributes(
  BackupId = "string",
  NeverExpires = TRUE|FALSE
)
}
}

\keyword{internal}
