% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_create_graph}
\alias{detective_create_graph}
\title{Creates a new behavior graph for the calling account, and sets that
account as the master account}
\usage{
detective_create_graph()
}
\description{
Creates a new behavior graph for the calling account, and sets that
account as the master account. This operation is called by the account
that is enabling Detective.
}
\details{
Before you try to enable Detective, make sure that your account has been
enrolled in Amazon GuardDuty for at least 48 hours. If you do not meet
this requirement, you cannot enable Detective. If you do meet the
GuardDuty prerequisite, then when you make the request to enable
Detective, it checks whether your data volume is within the Detective
quota. If it exceeds the quota, then you cannot enable Detective.

The operation also enables Detective for the calling account in the
currently selected Region. It returns the ARN of the new behavior graph.

\code{CreateGraph} triggers a process to create the corresponding data tables
for the new behavior graph.

An account can only be the master account for one behavior graph within
a Region. If the same account calls \code{CreateGraph} with the same master
account, it always returns the same behavior graph ARN. It does not
create a new behavior graph.
}
\keyword{internal}
