% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_get_members}
\alias{detective_get_members}
\title{Returns the membership details for specified member accounts for a
behavior graph}
\usage{
detective_get_members(GraphArn, AccountIds)
}
\arguments{
\item{GraphArn}{[required] The ARN of the behavior graph for which to request the member details.}

\item{AccountIds}{[required] The list of AWS account identifiers for the member account for which to
return member details.

You cannot use \code{GetMembers} to retrieve information about member
accounts that were removed from the behavior graph.}
}
\description{
Returns the membership details for specified member accounts for a
behavior graph.
}
\section{Request syntax}{
\preformatted{svc$get_members(
  GraphArn = "string",
  AccountIds = list(
    "string"
  )
)
}
}

\keyword{internal}
