% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_disassociate_admin_account}
\alias{fms_disassociate_admin_account}
\title{Disassociates the account that has been set as the AWS Firewall Manager
administrator account}
\usage{
fms_disassociate_admin_account()
}
\description{
Disassociates the account that has been set as the AWS Firewall Manager
administrator account. To set a different account as the administrator
account, you must submit an \code{AssociateAdminAccount} request.
}
\section{Request syntax}{
\preformatted{svc$disassociate_admin_account()
}
}

\keyword{internal}
