% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_describe_publishing_destination}
\alias{guardduty_describe_publishing_destination}
\title{Returns information about the publishing destination specified by the
provided destinationId}
\usage{
guardduty_describe_publishing_destination(DetectorId, DestinationId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector associated with the publishing destination
to retrieve.}

\item{DestinationId}{[required] The ID of the publishing destination to retrieve.}
}
\description{
Returns information about the publishing destination specified by the
provided \code{destinationId}.
}
\section{Request syntax}{
\preformatted{svc$describe_publishing_destination(
  DetectorId = "string",
  DestinationId = "string"
)
}
}

\keyword{internal}
