% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_accept_invitation}
\alias{macie2_accept_invitation}
\title{Accepts an Amazon Macie membership invitation that was received from a
specific account}
\usage{
macie2_accept_invitation(invitationId, masterAccount)
}
\arguments{
\item{invitationId}{[required] The unique identifier for the invitation to accept.}

\item{masterAccount}{[required] The AWS account ID for the account that sent the invitation.}
}
\description{
Accepts an Amazon Macie membership invitation that was received from a
specific account.
}
\section{Request syntax}{
\preformatted{svc$accept_invitation(
  invitationId = "string",
  masterAccount = "string"
)
}
}

\keyword{internal}
