% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_decline_invitations}
\alias{macie2_decline_invitations}
\title{Declines Amazon Macie membership invitations that were received from
specific accounts}
\usage{
macie2_decline_invitations(accountIds)
}
\arguments{
\item{accountIds}{[required] An array that lists AWS account IDs, one for each account that sent an
invitation to decline.}
}
\description{
Declines Amazon Macie membership invitations that were received from
specific accounts.
}
\section{Request syntax}{
\preformatted{svc$decline_invitations(
  accountIds = list(
    "string"
  )
)
}
}

\keyword{internal}
