% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_disassociate_from_master_account}
\alias{macie2_disassociate_from_master_account}
\title{Disassociates a member account from its Amazon Macie master account}
\usage{
macie2_disassociate_from_master_account()
}
\description{
Disassociates a member account from its Amazon Macie master account.
}
\section{Request syntax}{
\preformatted{svc$disassociate_from_master_account()
}
}

\keyword{internal}
