% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_bucket_statistics}
\alias{macie2_get_bucket_statistics}
\title{Retrieves (queries) aggregated statistical data for all the S3 buckets
that Amazon Macie monitors and analyzes}
\usage{
macie2_get_bucket_statistics(accountId)
}
\arguments{
\item{accountId}{The unique identifier for the AWS account.}
}
\description{
<p>Retrieves (queries) aggregated statistical data for all the S3 buckets that Amazon Macie monitors and analyzes.</p>
}
\section{Request syntax}{
\preformatted{svc$get_bucket_statistics(
  accountId = "string"
)
}
}

\keyword{internal}
