% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_master_account}
\alias{macie2_get_master_account}
\title{Retrieves information about the Amazon Macie master account for an
account}
\usage{
macie2_get_master_account()
}
\description{
Retrieves information about the Amazon Macie master account for an
account.
}
\section{Request syntax}{
\preformatted{svc$get_master_account()
}
}

\keyword{internal}
