% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_usage_statistics}
\alias{macie2_get_usage_statistics}
\title{Retrieves (queries) quotas and aggregated usage data for one or more
accounts}
\usage{
macie2_get_usage_statistics(filterBy, maxResults, nextToken, sortBy)
}
\arguments{
\item{filterBy}{An array of objects, one for each condition to use to filter the query
results. If the array contains more than one object, Amazon Macie uses
an AND operator to join the conditions specified by the objects.}

\item{maxResults}{The maximum number of items to include in each page of the response.}

\item{nextToken}{The nextToken string that specifies which page of results to return in a
paginated response.}

\item{sortBy}{The criteria to use to sort the query results.}
}
\description{
Retrieves (queries) quotas and aggregated usage data for one or more
accounts.
}
\section{Request syntax}{
\preformatted{svc$get_usage_statistics(
  filterBy = list(
    list(
      comparator = "GT"|"GTE"|"LT"|"LTE"|"EQ"|"NE"|"CONTAINS",
      key = "accountId"|"serviceLimit"|"freeTrialStartDate"|"total",
      values = list(
        "string"
      )
    )
  ),
  maxResults = 123,
  nextToken = "string",
  sortBy = list(
    key = "accountId"|"total"|"serviceLimitValue"|"freeTrialStartDate",
    orderBy = "ASC"|"DESC"
  )
)
}
}

\keyword{internal}
