% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_list_findings}
\alias{macie2_list_findings}
\title{Retrieves a subset of information about one or more findings}
\usage{
macie2_list_findings(findingCriteria, maxResults, nextToken,
  sortCriteria)
}
\arguments{
\item{findingCriteria}{The criteria to use to filter the results.}

\item{maxResults}{The maximum number of items to include in each page of the response.}

\item{nextToken}{The nextToken string that specifies which page of results to return in a
paginated response.}

\item{sortCriteria}{The criteria to use to sort the results.}
}
\description{
<p>Retrieves a subset of information about one or more findings.</p>
}
\section{Request syntax}{
\preformatted{svc$list_findings(
  findingCriteria = list(
    criterion = list(
      list(
        eq = list(
          "string"
        ),
        eqExactMatch = list(
          "string"
        ),
        gt = 123,
        gte = 123,
        lt = 123,
        lte = 123,
        neq = list(
          "string"
        )
      )
    )
  ),
  maxResults = 123,
  nextToken = "string",
  sortCriteria = list(
    attributeName = "string",
    orderBy = "ASC"|"DESC"
  )
)
}
}

\keyword{internal}
