% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_list_members}
\alias{macie2_list_members}
\title{Retrieves information about the accounts that are associated with an
Amazon Macie master account}
\usage{
macie2_list_members(maxResults, nextToken, onlyAssociated)
}
\arguments{
\item{maxResults}{The maximum number of items to include in each page of a paginated
response.}

\item{nextToken}{The nextToken string that specifies which page of results to return in a
paginated response.}

\item{onlyAssociated}{Specifies which accounts to include in the response, based on the status
of an account's relationship with the master account. By default, the
response includes only current member accounts. To include all accounts,
set the value for this parameter to false.}
}
\description{
Retrieves information about the accounts that are associated with an
Amazon Macie master account.
}
\section{Request syntax}{
\preformatted{svc$list_members(
  maxResults = 123,
  nextToken = "string",
  onlyAssociated = "string"
)
}
}

\keyword{internal}
