% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_list_tags_for_resource}
\alias{macie2_list_tags_for_resource}
\title{Retrieves the tags (keys and values) that are associated with a
classification job, custom data identifier, findings filter, or member
account}
\usage{
macie2_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the classification job, custom data
identifier, findings filter, or member account.}
}
\description{
Retrieves the tags (keys and values) that are associated with a
classification job, custom data identifier, findings filter, or member
account.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  resourceArn = "string"
)
}
}

\keyword{internal}
