% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_test_custom_data_identifier}
\alias{macie2_test_custom_data_identifier}
\title{Tests a custom data identifier}
\usage{
macie2_test_custom_data_identifier(ignoreWords, keywords,
  maximumMatchDistance, regex, sampleText)
}
\arguments{
\item{ignoreWords}{An array that lists specific character sequences (ignore words) to
exclude from the results. If the text matched by the regular expression
is the same as any string in this array, Amazon Macie ignores it. The
array can contain as many as 10 ignore words. Each ignore word can
contain 4 - 90 characters. Ignore words are case sensitive.}

\item{keywords}{An array that lists specific character sequences (keywords), one of
which must be within proximity (maximumMatchDistance) of the regular
expression to match. The array can contain as many as 50 keywords. Each
keyword can contain 4 - 90 characters. Keywords aren't case sensitive.}

\item{maximumMatchDistance}{The maximum number of characters that can exist between text that
matches the regex pattern and the character sequences specified by the
keywords array. Macie includes or excludes a result based on the
proximity of a keyword to text that matches the regex pattern. The
distance can be 1 - 300 characters. The default value is 50.}

\item{regex}{[required] The regular expression (\emph{regex}) that defines the pattern to match. The
expression can contain as many as 512 characters.}

\item{sampleText}{[required] The sample text to inspect by using the custom data identifier. The text
can contain as many as 1,000 characters.}
}
\description{
Tests a custom data identifier.
}
\section{Request syntax}{
\preformatted{svc$test_custom_data_identifier(
  ignoreWords = list(
    "string"
  ),
  keywords = list(
    "string"
  ),
  maximumMatchDistance = 123,
  regex = "string",
  sampleText = "string"
)
}
}

\keyword{internal}
