% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_untag_resource}
\alias{macie2_untag_resource}
\title{Removes one or more tags (keys and values) from a classification job,
custom data identifier, findings filter, or member account}
\usage{
macie2_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the classification job, custom data
identifier, findings filter, or member account.}

\item{tagKeys}{[required] The key of the tag to remove from the resource. To remove multiple tags,
append the tagKeys parameter and argument for each additional tag to
remove, separated by an ampersand (&).}
}
\description{
Removes one or more tags (keys and values) from a classification job,
custom data identifier, findings filter, or member account.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
