% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_update_organization_configuration}
\alias{macie2_update_organization_configuration}
\title{Updates the Amazon Macie configuration settings for an AWS organization}
\usage{
macie2_update_organization_configuration(autoEnable)
}
\arguments{
\item{autoEnable}{[required] Specifies whether Amazon Macie is enabled automatically for each
account, when the account is added to the AWS organization.}
}
\description{
Updates the Amazon Macie configuration settings for an AWS organization.
}
\section{Request syntax}{
\preformatted{svc$update_organization_configuration(
  autoEnable = TRUE|FALSE
)
}
}

\keyword{internal}
