% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_associate_subnets}
\alias{networkfirewall_associate_subnets}
\title{Associates the specified subnets in the Amazon VPC to the firewall}
\usage{
networkfirewall_associate_subnets(UpdateToken, FirewallArn,
  FirewallName, SubnetMappings)
}
\arguments{
\item{UpdateToken}{An optional token that you can use for optimistic locking. Network
Firewall returns a token to your requests that access the firewall. The
token marks the state of the firewall resource at the time of the
request.

To make an unconditional change to the firewall, omit the token in your
update request. Without the token, Network Firewall performs your
updates regardless of whether the firewall has changed since you last
retrieved it.

To make a conditional change to the firewall, provide the token in your
update request. Network Firewall uses the token to ensure that the
firewall hasn't changed since you last retrieved it. If it has changed,
the operation fails with an \code{InvalidTokenException}. If this happens,
retrieve the firewall again to get a current copy of it with a new
token. Reapply your changes as needed, then try the operation again
using the new token.}

\item{FirewallArn}{The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.}

\item{FirewallName}{The descriptive name of the firewall. You can't change the name of a
firewall after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{SubnetMappings}{[required] The IDs of the subnets that you want to associate with the firewall.}
}
\description{
Associates the specified subnets in the Amazon VPC to the firewall. You
can specify one subnet for each of the Availability Zones that the VPC
spans.
}
\details{
This request creates an AWS Network Firewall firewall endpoint in each
of the subnets. To enable the firewall's protections, you must also
modify the VPC's route tables for each subnet's Availability Zone, to
redirect the traffic that's coming into and going out of the zone
through the firewall endpoint.
}
\section{Request syntax}{
\preformatted{svc$associate_subnets(
  UpdateToken = "string",
  FirewallArn = "string",
  FirewallName = "string",
  SubnetMappings = list(
    list(
      SubnetId = "string"
    )
  )
)
}
}

\keyword{internal}
