% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_tag_resource}
\alias{networkfirewall_tag_resource}
\title{Adds the specified tags to the specified resource}
\usage{
networkfirewall_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{Tags}{[required]}
}
\description{
Adds the specified tags to the specified resource. Tags are key:value
pairs that you can use to categorize and manage your resources, for
purposes like billing. For example, you might set the tag key to
"customer" and the value to the customer name or ID. You can specify one
or more tags to add to each AWS resource, up to 50 tags for a resource.
}
\details{
You can tag the AWS resources that you manage through AWS Network
Firewall: firewalls, firewall policies, and rule groups.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceArn = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
