% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram_operations.R
\name{ram_list_pending_invitation_resources}
\alias{ram_list_pending_invitation_resources}
\title{Lists the resources in a resource share that is shared with you but that
the invitation is still pending for}
\usage{
ram_list_pending_invitation_resources(resourceShareInvitationArn,
  nextToken, maxResults)
}
\arguments{
\item{resourceShareInvitationArn}{[required] The Amazon Resource Name (ARN) of the invitation.}

\item{nextToken}{The token for the next page of results.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}
}
\description{
Lists the resources in a resource share that is shared with you but that
the invitation is still pending for.
}
\section{Request syntax}{
\preformatted{svc$list_pending_invitation_resources(
  resourceShareInvitationArn = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
