% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_describe_attack}
\alias{shield_describe_attack}
\title{Describes the details of a DDoS attack}
\usage{
shield_describe_attack(AttackId)
}
\arguments{
\item{AttackId}{[required] The unique identifier (ID) for the attack that to be described.}
}
\description{
Describes the details of a DDoS attack.
}
\section{Request syntax}{
\preformatted{svc$describe_attack(
  AttackId = "string"
)
}
}

\keyword{internal}
