% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_describe_drt_access}
\alias{shield_describe_drt_access}
\title{Returns the current role and list of Amazon S3 log buckets used by the
DDoS Response Team (DRT) to access your AWS account while assisting with
attack mitigation}
\usage{
shield_describe_drt_access()
}
\description{
Returns the current role and list of Amazon S3 log buckets used by the
DDoS Response Team (DRT) to access your AWS account while assisting with
attack mitigation.
}
\section{Request syntax}{
\preformatted{svc$describe_drt_access()
}
}

\keyword{internal}
