% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_list_sql_injection_match_sets}
\alias{wafregional_list_sql_injection_match_sets}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_list_sql_injection_match_sets(NextMarker, Limit)
}
\arguments{
\item{NextMarker}{If you specify a value for \code{Limit} and you have more
SqlInjectionMatchSet objects than the value of \code{Limit}, AWS WAF returns
a \code{NextMarker} value in the response that allows you to list another
group of \code{SqlInjectionMatchSets}. For the second and subsequent
\code{ListSqlInjectionMatchSets} requests, specify the value of \code{NextMarker}
from the previous response to get information about another batch of
\code{SqlInjectionMatchSets}.}

\item{Limit}{Specifies the number of SqlInjectionMatchSet objects that you want AWS
WAF to return for this request. If you have more \code{SqlInjectionMatchSet}
objects than the number you specify for \code{Limit}, the response includes a
\code{NextMarker} value that you can use to get another batch of \code{Rules}.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see
\href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic}
in the developer guide.
}
\details{
\strong{For the latest version of AWS WAF}, use the AWS WAFV2 API and see the
\href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
With the latest version, AWS WAF has a single set of endpoints for
regional and global use.

Returns an array of SqlInjectionMatchSet objects.
}
\section{Request syntax}{
\preformatted{svc$list_sql_injection_match_sets(
  NextMarker = "string",
  Limit = 123
)
}
}

\examples{
\dontrun{
# The following example returns an array of up to 100 SQL injection match
# sets.
svc$list_sql_injection_match_sets(
  Limit = 100L
)
}

}
\keyword{internal}
