% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_list_resources_for_web_acl}
\alias{wafv2_list_resources_for_web_acl}
\title{This is the latest version of AWS WAF, named AWS WAFV2, released in
November, 2019}
\usage{
wafv2_list_resources_for_web_acl(WebACLArn, ResourceType)
}
\arguments{
\item{WebACLArn}{[required] The Amazon Resource Name (ARN) of the Web ACL.}

\item{ResourceType}{Used for web ACLs that are scoped for regional applications. A regional
application can be an Application Load Balancer (ALB), an API Gateway
REST API, or an AppSync GraphQL API.}
}
\description{
This is the latest version of \strong{AWS WAF}, named AWS WAFV2, released in
November, 2019. For information, including how to migrate your AWS WAF
resources from the prior release, see the \href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
}
\details{
Retrieves an array of the Amazon Resource Names (ARNs) for the regional
resources that are associated with the specified web ACL. If you want
the list of AWS CloudFront resources, use the AWS CloudFront call
\code{ListDistributionsByWebACLId}.
}
\section{Request syntax}{
\preformatted{svc$list_resources_for_web_acl(
  WebACLArn = "string",
  ResourceType = "APPLICATION_LOAD_BALANCER"|"API_GATEWAY"|"APPSYNC"
)
}
}

\keyword{internal}
